fasta_slicer/requirements.txt
biopython==1.68
numpy==1.11.2
pandas==0.19.1
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
